🪲 Print kernel info for debugging
+ Created a new module for debugging purposes. Added pp_kernel_function.py which will allow the programmer to view information about the kernel function and the arguments passed in a tabular format

+ pp_table_generator.py : Helper function to easily create pretty-tables

+++ python/seastar/code_gen/cuda_python/cuda_helper.py
@@ -80,7 +80,8 @@ def copy_arguments_to_gpu(argument_list, stream):
 
     # copying graph vector arguments to GPU Device
     for argument in graph_vector_arguments:
-        host_argument, host_argument_size = createNpArray(argument.flatten().tolist(), np.float64)
+        # breakpoint()
+        host_argument, host_argument_size = createNpArray(argument.flatten().tolist(), np.float32)
         device_argument, argument_class = allocAndCopyToDevice(host_argument, host_argument_size, stream)
         kernel_arguments.append(device_argument)
 
@@ -90,13 +91,14 @@ def copy_arguments_to_gpu(argument_list, stream):
 
     # copying graph csr arguments to GPU Device
     for argument in graph_csr_arguments:
-        host_argument, host_argument_size = createNpArray(argument.flatten().tolist(), np.int64)
+        # breakpoint()
+        host_argument, host_argument_size = createNpArray(argument.flatten().tolist(), np.int32)
         device_argument, argument_class = allocAndCopyToDevice(host_argument, host_argument_size, stream)
         kernel_arguments.append(device_argument)
 
     # copying scalar arguments to GPU Device
     for argument in scalar_arguments:
-        host_argument, host_argument_size = createNpArray([argument], np.int64)
+        host_argument, host_argument_size = createNpArray([argument], np.int32)
         kernel_arguments.append(host_argument)
 
     # converting it into a form that can be passed
